Wrote Mapper and Reducer for testGen
These scripts take the raw JSON data as input to the mapper and produce
a JSON file of query data. Each entry in the output represents a single query.
These entries hold 3 fields: the shown items of this query (total over all shown
pages), the clicks the user has clicked on prior to the search, and the items
from this query that the user ends up clicking on. These fields are each arrays.

Should be used as follows:
> testGenMapper.py 000000_0 | sort | testGenReducer.py > results
and results holds the data in JSON format.